handle history writing internally
The problem with ->WriteHistory() is that it outputs the history entires
unmodified, which for multiline entries results in them being turned
into multiple entries when you restart tdsql.

This version uses YAML for the history file, backward-compatibly (checks
for ---\n at the top), and while we're loading YAML, also uses it to
load the rc file.